Add files via upload
Generative Adversarial Networks (GANs) are quite interesting which can be used to generative new data by learning from existing data. This is a game between two neural networks called generative network and discriminative network, which explains the ‘Adversarial’ part of it. The purpose of this game is to favor the generative network and say it should be the winner, although this doesn’t mean we should be bias. The reason is when the generator wins we can be confident that it generator very realistic data. For this task, a specialized model is used which is specidically designed for image data, called Deep Convolutional Generative Adversarial Network (DCGAN). DCGANs have the convolutional layer used for extracting high-level features like edges from input images